requirements.txt
streamlit>=1.27.0
tensorflow>=2.12.0
torch>=2.1.0
transformers>=4.43.0
Pillow>=9.0.0
numpy>=1.26.0
matplotlib>=3.8.1
pandas>=2.1.0
scipy>=1.11.0
requests>=2.31.0